pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "E-Commerce test project"
authors = [
    {name = "Bogdan Borutenko",email = "bogdanborutenko@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=24.10.0,<25.0.0)",
    "isort (>=5.13.2,<6.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-asyncio (>=0.25.2,<0.26.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi (>=0.115.6,<0.115.7)",
    "python-json-logger (>=3.2.1,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py39']


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = ["test_*.py", "*_test.py"]